now <BuscarAssistido/> is a module that receives a callback
